+++ shared/src/contactEstablishmentRequest/contactEstablishmentRequest.dto.ts
@@ -2,7 +2,7 @@ import { WithAcquisition } from "../acquisition.dto";
 import { LocationId } from "../address/address.dto";
 import { ImmersionObjective } from "../convention/convention.dto";
 import { DiscussionId } from "../discussion/discussion.dto";
-import { ContactMethod } from "../formEstablishment/FormEstablishments.dto";
+import { ContactMethod } from "../formEstablishment/formEstablishment.dto";
 import { AppellationCode } from "../romeAndAppellationDtos/romeAndAppellation.dto";
 import { SiretDto } from "../siret/siret";
