+++ shared/src/search/SearchResult.dto.ts
@@ -1,5 +1,5 @@
 import { AddressDto, LocationId } from "../address/address.dto";
-import { ContactMethod } from "../formEstablishment/FormEstablishments.dto";
+import { ContactMethod } from "../formEstablishment/formEstablishment.dto";
 import { GeoPositionDto } from "../geoPosition/geoPosition.dto";
 import {
   AppellationDto,
